mymodules/aisight/requirements.txt
python-dateutil
tqdm
pandas
numpy
pyproj
shapely
osgeo